requirements.txt
google-generativeai
python-dotenv
ddgs